ManuOptimizer/ManuOptimizer-backend/requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
flask_sqlalchemy==3.1.1
PuLP==2.8.0
python-dotenv==0.19.0
waitress==2.1.2
